Move and rename `ExecutionOptions`
